\name{contFracR-package}
\alias{contFracR-package}
\alias{contFracR}
\docType{package}
\title{ Functions to generate and evaluate continued fractions}
\description{
This package includes functions to convert a rational number to its continued fraction form, to convert any defined continued fraction to a numeric value, and to generate estimates of some common irrationals (e, pi, phi) via continued fraction representations. There's also a function which  generates LaTeX code to draw the continued fraction itself. 
 
}

\author{
Carl Witthoft, with attributions as noted in the individual help pages

Maintainer:Carl Witthoft carl@witthoft.com

}

